Removing the stashCard file and renaming streamCard to ArticleCard.  Article card is now dynamic and will render either a card for a stream article or a stash article based on the URL.
